+++ config.h
@@ -74,6 +74,7 @@ struct config {
     struct {
         enum cursor_style style;
         bool blink;
+        bool hide_when_typing;
         struct {
             uint32_t text;
             uint32_t cursor;